+++ cpg_utils/workflows/filetypes.py
@@ -9,7 +9,7 @@ from typing import Union
 from hailtop.batch import ResourceGroup, ResourceFile, Batch
 
 from cpg_utils import Path, to_path
-from cpg_utils.flows.utils import exists
+from cpg_utils.workflows.utils import exists
 
 
 class AlignmentInput(ABC):
